# docs/openapi.yaml
openapi: "3.1.0"
...
info:
  title: Chronos Ledger API
  version: "0.10.0" # x-release-please-version
  description: |
    Organization Schedule & Attendance Management REST API.

    **Authentication**: All endpoints except `/config` and `/auth/login`
    require a Bearer JWT in the `Authorization` header. An API key in
    `X-API-Key` is accepted in place of one anywhere, which is how a lobby
    terminal with no one to sign in reaches `/guest/register-checkin` and
    `/guest/directory`. A key can be narrowed to part of the API and given
    an end date: see the `ApiKeyAuth` security scheme.

    **Roles**: `SUPER_ADMIN` > `UNIT_ADMIN` > `STAFF` > `MEMBER`.
    Role requirements are noted on each endpoint.
  contact:
    name: Chronos Ledger Contributors
    email: admin@org.internal
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost
    description: Local development (Docker Compose)
  - url: http://{server_ip}
    description: Organization intranet
    variables:
      server_ip:
        default: "192.168.1.100"

# ─────────────────────────────────────────────
# Reusable components
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A machine credential bound to a service-account user. Accepted anywhere
        a Bearer token is, and acts as the bound user, so roles and unit scoping
        apply unchanged. This is how a lobby kiosk authenticates.

        A key can also be narrowed and given an end date, which a Bearer token
        cannot be. `scopes` is a comma separated list of `<area>:read` and
        `<area>:write`, where the area is the path segment after `/api/v1`:
        `schedule`, `resources`, `attendance`, `users`, `ingestion`, `guest`,
        `sync`, `config`, `auth`, `api-keys`. `GET`, `HEAD` and `OPTIONS` are
        reads and everything else is a write, so `schedule:read` passes
        `GET /schedule/slots` and is refused `POST /schedule/slots` with a 403
        naming `schedule:write`. The single scope `*` is every area, which is
        what a key issued before scopes existed holds and what a key created
        without naming any gets.

        `expires_at` is checked on the request, not by a sweep, so a key stops
        working the moment it runs out and answers 401 `API key expired`. A
        null expiry never runs out.

        The first-login gate does not apply to a key. It holds an admin
        account until the password it was handed is changed, and a key never
        uses that password, so a service account at an admin role works
        through its key without anyone signing in as it.
  schemas:
...
    InstitutionalRole:
      type: string
      enum: [SUPER_ADMIN, UNIT_ADMIN, STAFF, MEMBER]
      description: Role assigned to a system user.
    AccessReadiness:
      type: string
      enum: [OPEN_AD_HOC, BUSY, CRITICAL_DO_NOT_DISTURB, VERY_FREE]
      description: Staff availability / occupancy index visible to members.
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: staff@org.internal
        password:
          type: string
          format: password
          example: "••••••••"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: "Signed JWT. Include as `Authorization: Bearer <token>`."
        token_type:
          type: string
          example: bearer
        user_id:
          type: string
          example: FAC001
        role:
          $ref: "#/components/schemas/InstitutionalRole"
        full_name:
          type: string
          example: Dr. Priya Sharma
        initial_login_state:
          type: boolean
          description: True on first login, client should prompt password change.
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
          minLength: 12
          description: >
            At least PASSWORD_MIN_LENGTH characters, 12 unless the deployment
            has raised it, and not a placeholder published in this repository.
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        email_address:
          type: string
        role_type:
          $ref: "#/components/schemas/InstitutionalRole"
        unit_code:
          type: string
          nullable: true
        assigned_base_station:
          type: string
          nullable: true
        current_occupancy_index:
          $ref: "#/components/schemas/AccessReadiness"
        reporting_line_manager:
          type: string
          nullable: true
        deactivated_at:
          type: string
          format: date-time
          nullable: true
          description: When the account was deactivated, or null while it is active.

    # ── API keys ─────────────────────────────
    ApiKeyCreate:
      type: object
      required: [label, user_id]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 100
          description: What this key is for, so it can be recognised in the list.
          example: PulseWard HMS
        user_id:
          type: string
          description: >
            The user this key acts as. The key can never do more than that
            user can, whatever its scopes say.
          example: SVC001
        scopes:
          type: array
          nullable: true
          minItems: 1
          items:
            type: string
            pattern: "^(\\*|[a-z][a-z-]*:(read|write))$"
          description: >
            Areas this key may reach. Omit the field for an unrestricted key;
            an empty array is refused, since a key that can reach nothing is
            not a key anybody wants. Stored sorted and deduplicated.
          example: [schedule:read, resources:read]
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: >
            When the key stops working. Omit for a key that never does. A time
            already past is refused at creation.
          example: "2027-01-01T00:00:00Z"
    ApiKeyResponse:
      type: object
      properties:
        id:
          type: integer
        key_prefix:
          type: string
          description: The first 12 characters, enough to tell two keys apart.
          example: ck_8Kd2mQ7x
        label:
          type: string
        user_id:
          type: string
        scopes:
          type: string
          description: Comma separated, or `*` for a key that was never narrowed.
          example: resources:read,schedule:read
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
    ApiKeyCreated:
      allOf:
        - $ref: "#/components/schemas/ApiKeyResponse"
        - type: object
          required: [api_key]
          properties:
            api_key:
              type: string
              description: >
                The raw key. Returned exactly once, here. Only its hash is
                stored, so it cannot be shown again and a lost key has to be
                revoked and reissued.
              example: ck_8Kd2mQ7xR3nL9vB5tY1wZ0aC6eF4gH2jK8mN

    # ── Resources ─────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Not authorized.
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but insufficient role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Validation error (e.g. malformed request body).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: >
        The caller is past its budget for this route. Sized by the
        RATE_LIMIT_* settings; the counters live in Redis, and an instance
        that cannot reach Redis stops applying the limits rather than
        refusing the requests.
      headers:
        Retry-After:
          description: Seconds until the budget refills.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

# Either credential is accepted on every endpoint below.
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login and receive JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Exchange a refresh token for a new access token
      description: >
        A refresh token works exactly once. The response carries its
        replacement, so a client that loses the response signs in again. A
        used token presented again more than ten seconds after its use ends
        every token descended from the same sign-in, on the reading that two
        parties hold it; inside those ten seconds it is refused and nothing
        else changes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: A new access token and a replacement refresh token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: >
            The token is unknown, already used, expired, or belongs to an
            account that no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke a refresh token
      description: >
        Deletes the refresh token named in the body along with the used tokens
        it replaced, which ends that sign-in. An unknown token is not an
        error, so a client that has already lost its token can call this
        without handling a failure. The access token keeps working until it
        expires: there is no server-side revocation of access tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token revoked, or it was already gone.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: User profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/change-password:
    post:
      tags: [Authentication]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: >
            Password changed. Every existing session is revoked and the
            calendar feed token is rotated. `refresh_token` replaces the one
            the caller holds; without storing it the caller is signed out when
            its access token expires.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated successfully.
                  refresh_token:
                    type: string
        "400":
          description: >
            Current password incorrect, or the new password is the one the
            account already has. A password that breaks the length or
            placeholder rule is a 422 instead.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Users ─────────────────────────────────
...
  /api/v1/api-keys/:
    get:
      tags: [API Keys]
      summary: List issued keys
      description: >
        Requires `SUPER_ADMIN`. Shows what each key is scoped to and when it
        runs out, which is how an unrestricted key left over from before
        scopes existed becomes visible.
      responses:
        "200":
          description: Every key, without its secret.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [API Keys]
      summary: Issue a key
      description: >
        Requires `SUPER_ADMIN`. The raw key is in the response and nowhere
        else. Bind it to a service account rather than to a person: the key
        acts as that user, so a person's key inherits everything they can do.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreate"
      responses:
        "201":
          description: Key issued. The `api_key` field will not be shown again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such user to bind the key to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The user to bind the key to has been deactivated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: >
            A scope that is not a scope, an empty scope list, or an
            `expires_at` already in the past.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/api-keys/{key_id}:
    delete:
      tags: [API Keys]
      summary: Revoke a key
      description: >
        Requires `SUPER_ADMIN`. Takes effect on the next request that uses it.
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Resources ─────────────────────────────
...